% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CResource.R
\name{add.consumed}
\alias{add.consumed}
\title{Function to increase the consumed resources}
\usage{
add.consumed(resource, t = 0, ev = 0, it = 0)
}
\arguments{
\item{resource}{Object of class \code{\link{CResource}} to which the consumption will be added}

\item{t}{Time to add}

\item{ev}{Number of evaluations to add}

\item{it}{Number of iterations to add}
}
\value{
A new object of class \code{\link{CResource}} with the updated consumed resources
}
\description{
This function increases the consumed resources
}

