% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CResource.R
\name{addConsumed}
\alias{addConsumed}
\title{Function to increase the consumed resources}
\usage{
addConsumed(resource, t = 0, ev = 0, it = 0)
}
\arguments{
\item{resource}{Object of class \code{\linkS4class{CResource}} to which the consumption will be added}

\item{t}{Time to add}

\item{ev}{Number of evaluations to add}

\item{it}{Number of iterations to add}
}
\value{
This method does not return any output. For efficiecy reasons the object passed is modified. This can lead to problems if not handled properly, so use with caution!
}
\description{
This function increases the consumed resources
}

