% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MHResult.R
\name{algorithm}
\alias{algorithm}
\title{Accession function for the \code{algorithm} slot of the object}
\usage{
algorithm(result)
}
\arguments{
\item{result}{Object of class \code{\linkS4class{MHResult}} whose information we are quering}
}
\value{
Content of the \code{algorithm} slot
}
\description{
Accession function for the \code{algorithm} slot of the object
}

