% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/acceptance_functions.R
\name{always.accept}
\alias{always.accept}
\title{Dummy function to always accept changes.}
\usage{
always.accept(delta, ...)
}
\arguments{
\item{delta}{Increment in the evaluation function computed as the evaluation of the new solution minus the evaluation of the existing solution.}
}
\value{
This function always accepts the new solution, so it returns always \code{TRUE}.
}
\description{
This function returns \code{TRUE}, regardless the increment in the evaluation function.
}
\examples{
always.accept (-1)
always.accept (1)
}
\seealso{
Other Acceptance functions.: \code{\link{boltzmann.accept}};
  \code{\link{threshold.accept}}
}

