% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/acceptance_functions.R
\name{alwaysAccept}
\alias{alwaysAccept}
\title{Dummy function to always accept changes.}
\usage{
alwaysAccept(delta, ...)
}
\arguments{
\item{delta}{Increment in the evaluation function computed as the evaluation of the new solution minus the evaluation of the existing solution.}
}
\value{
This function always accepts the new solution, so it returns always \code{TRUE}.
}
\description{
This function returns \code{TRUE}, regardless the increment in the evaluation function.
}
\examples{
alwaysAccept (-1)
alwaysAccept (1)
}
\seealso{
Other Acceptance functions.: \code{\link{boltzmannAccept}};
  \code{\link{thresholdAccept}}
}

