% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_search.R
\name{basicLocalSearch}
\alias{basicLocalSearch}
\title{Basic local search}
\usage{
basicLocalSearch(evaluate, initial.solution, neighborhood, selector,
  do.log = TRUE, save.sols = FALSE, verbose = TRUE,
  non.valid = "ignore", valid = allValid, correct = doNothing,
  resources = cResource(), ...)
}
\arguments{
\item{evaluate}{Function of a single parameter. Given a solution, the function returns its evaluation as a real number}

\item{initial.solution}{Solution from where the search will start}

\item{neighborhood}{Object representing the type of neighborhood to be used}

\item{selector}{A function used to select a solution from the neighborhood. For an example of the parameters required and the result it should produce, see \code{\link{greedySelector}}}

\item{do.log}{Logic value to indicate whether the progress shoul be tracked or not}

\item{save.sols}{Logic value to indicate whether the evaluated solutions should be stored}

\item{verbose}{Logic value to indicate whether the function should print information about the search}

\item{non.valid}{Action to be performed when a non valid solution is considered. The options are \code{'ignore'}, meaning that the solution is considered anyway, \code{'discard'}, meaning that the solution is not considered and \code{'correct'}, meaning that the solution has to be corrected. This parameter has to be set only when there can be non valid solutions}

\item{valid}{A function that, given a solution, determines whether it is valid or not}

\item{correct}{A function that, given a non valid solution, corrects it. This optional parameter has to be set only with the \code{'correct'} option}

\item{resources}{Object of class \code{\linkS4class{cResource}} representing the available computational resources for the search}

\item{...}{Special argument to pass additional parameters to the functions used in the search}
}
\value{
The function returns an object of class \code{\linkS4class{mHResult}} with all the information about the search
}
\description{
\code{basicLocalSearch} performs a simple local search and returns the best solution
}

