% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/swarm.R
\name{basicPso}
\alias{basicPso}
\title{Basic PSO}
\usage{
basicPso(evaluate, initial.positions, initial.velocity, max.velocity,
  c.personal, c.best, do.log = TRUE, verbose = TRUE, non.valid = "ignore",
  valid = allValid, correct = doNothing, resources = cResource(), ...)
}
\arguments{
\item{evaluate}{Function of a single parameter. Given a solution, the function returns its evaluation as a real number.}

\item{initial.positions}{A list of solutions representing the initial positions of the particles}

\item{initial.velocity}{A list containing the intial velocities of the particles}

\item{max.velocity}{Maximum velocity allowed to a particle}

\item{c.personal}{Coefficient for the personal best solution (i.e., the cognitive component)}

\item{c.best}{Coefficient for the global best solution (i.e., the social component)}

\item{do.log}{Logic value to indicate whether the progress shoul be tracked or not}

\item{verbose}{Logic value to indicate whether the function should print information about the search}

\item{non.valid}{Action to be performed when a non valid solution is considered. The options are \code{'ignore'}, meaning that the solution is considered anyway, \code{'discard'}, meaning that the solution is not considered and \code{'correct'}, meaning that the solution has to be corrected. This parameter has to be set only when there can be non valid solutions}

\item{valid}{A function that, given a solution, determines whether it is valid or not}

\item{correct}{A function that, given a non valid solution, corrects it. This optional parameter has to be set only with the \code{'correct'} option}

\item{resources}{Object of class \code{\linkS4class{cResource}} representing the available computational resources for the search. Bear in mind that there is no other stop criterion beyond a limited amount of resources. Therefore, you should set, at least, a limit to the total time, evaluations or iterations}

\item{...}{Special argument to pass additional parameters to the functions used in the search}
}
\value{
The function returns an object of class \code{\linkS4class{mHResult}} with all the information about the search
}
\description{
This function implements a basic version of the PSO algorithm
}
\details{
This basic version of the PSO only works with numeric problems and, thus, the solutions have to be numeric vectors.
}
\seealso{
\code{\link{tournamentSelection}}

Other Swarm Intelligence: \code{\link{basicAco}}
}

