% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianNetwork.R
\docType{class}
\name{BayesianNetwork-class}
\alias{BayesianNetwork-class}
\title{An S4 class to represent distributions based on Bayesian networks}
\description{
An S4 class to represent distributions based on Bayesian networks
}
\section{Slots}{

\describe{
\item{\code{network}}{An object of the type bn.fit (see bnLearn pacakge for details)}

\item{\code{factor.population}}{A logical value which determine if the learned population was factor or not.}
}}

