% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianNetwork.R
\name{bayesianNetwork}
\alias{bayesianNetwork}
\title{Constructor of EBNA model}
\usage{
bayesianNetwork(data, ...)
}
\arguments{
\item{data}{The dataframe containing the initial-population to construct the bayesian network}

\item{...}{Ignored}
}
\value{
An object of class \code{\linkS4class{BayesianNetwork}} that includes the Bayesian Network of the given data
}
\description{
This function creates an object of class \code{\linkS4class{BayesianNetwork}}
}
\seealso{
Other EDA: \code{\link{ehbsa}}, \code{\link{mallowsModel}},
  \code{\link{plackettLuce}},
  \code{\link{univariateMarginals}}
}

