% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianNetwork.R
\name{bin2cat}
\alias{bin2cat}
\title{Converts a vector from logical to categorical.}
\usage{
bin2cat(bin.v)
}
\arguments{
\item{bin.v}{logical vector to be converted.}
}
\value{
a categorical vector
}
\description{
Converts a vector from logical to categorical.
}

