% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mutation.R
\name{binary.mutation}
\alias{binary.mutation}
\title{Binary mutation}
\usage{
binary.mutation(solution, ratio, ...)
}
\arguments{
\item{solution}{A logical vector to be mutated.}

\item{ratio}{Ratio of the positions to be mutated. It has to be a value greater than 0.}
}
\description{
This function takes random positions of a logical vector and changes their value
}
\examples{
F.sol <- rep(FALSE , 10)
binary.mutation (F.sol , ratio = 0.1)
binary.mutation (F.sol , ratio = 0.1)
binary.mutation (F.sol , ratio = 0.5)
}

