% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation.R
\name{binaryMutation}
\alias{binaryMutation}
\title{Binary mutation}
\usage{
binaryMutation(solution, ratio, ...)
}
\arguments{
\item{solution}{A logical vector to be mutated.}

\item{ratio}{Ratio of the positions to be mutated. It has to be a value greater than 0.}
}
\description{
This function takes random positions of a logical vector and changes their value
}
\examples{

F.sol <- rep(FALSE, 10)
binaryMutation(F.sol, ratio=0.1)
binaryMutation(F.sol, ratio=0.1)
binaryMutation(F.sol, ratio=0.5)

}

