% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/acceptance_functions.R
\name{boltzmann.accept}
\alias{boltzmann.accept}
\title{Probabilistic acceptance based on Boltzmann's distribution.}
\usage{
boltzmann.accept(delta, temperature = 1, ...)
}
\arguments{
\item{delta}{Increment in the evaluation function computed as the evaluation of the new solution minus the evaluation of the existing solution.}

\item{temperature}{Temperature parameter in Boltzmann's distribution. This temperature has to be a value strictly greater than 0. By default this parameter is equal to 1.}
}
\value{
\code{TRUE} with probability given by exp(-1*delta/tempreature).
}
\description{
This function accepts solutions (returns \code{TRUE}) with probability given by Boltzmann's distribution.
}
\examples{
boltzmann.accept (delta = -1 , temperature = 2)
boltzmann.accept (delta = -3 , th = -2)
boltzmann.accept (delta = -1 , th = 0)
}
\seealso{
Other Acceptance functions.: \code{\link{always.accept}};
  \code{\link{threshold.accept}}
}

