% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Pheromone.R
\name{build.solution}
\alias{build.solution}
\title{Function to build solutions according to the pheromone trail}
\usage{
build.solution(object, n = 1, seed = NULL, ...)
}
\arguments{
\item{object}{Object of class \code{\linkS4class{VectorPheromone}} representing the pheromone model to be updated}

\item{n}{Numeric value indicating the amount of solutions to be created}

\item{seed}{Numeric value to indicate the seed used for the random number generator. If \code{NULL}, the seed is not set}
}
\description{
This function creates new solutions using the pheromone trail
}

