% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianNetwork.R, R/PlackettLuce.R
\name{calc.prob}
\alias{calc.prob}
\title{Calculates the probabilities of the given model for the especified population.}
\usage{
calc.prob(model, data)

calc.prob(model, data)
}
\arguments{
\item{model}{object of the class PlackettLuce.}

\item{data}{population of individuals whoose probabilities are desired to know.}

\item{object}{object of the class BayesianNetwork.}

\item{data}{population whoose probabilities are desired to know.}
}
\value{
a vector containing the probabilities.
}
\description{
Calculates the probabilities of the given model for the especified population.

Calculates the probabilities of the given model for the population especified.
}

