% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianNetwork.R
\name{cat2bin}
\alias{cat2bin}
\title{Converts a vector from categorical to logical.}
\usage{
cat2bin(catVector)
}
\arguments{
\item{catVector}{categorical vector to be converted.}
}
\value{
a logical vector
}
\description{
Converts a vector from categorical to logical.
}

