% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/evolutionary.R
\name{check.solutions}
\alias{check.solutions}
\title{Chek solutions}
\usage{
check.solutions(solution, non.valid, valid, correct)
}
\arguments{
\item{solution}{The solution to be cheked}

\item{non.valid}{Action to be performed if the solution is not valid. Options are \code{'ignore'},\code{'discard'} and \code{'correct'}.}

\item{valid}{Function to check the validity of the solution}

\item{correct}{Function to correct the solution.}
}
\description{
This function checks the validity of a solution and does the required action if the soution is not valid
}
\keyword{internal}

