% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CResource.R
\name{consumed.iterations}
\alias{consumed.iterations}
\title{Function to get the number of iterations consumed}
\usage{
consumed.iterations(resource)
}
\arguments{
\item{resource}{Object of class \code{\linkS4class{CResource}} whose information we are quering}
}
\value{
Number of iterations consumed
}
\description{
This function returns the number of iterations consumed so far
}

