% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CResource.R
\name{consumed.time}
\alias{consumed.time}
\title{Function to get the consumed time}
\usage{
consumed.time(resource)
}
\arguments{
\item{resource}{Object of class \code{\linkS4class{CResource}} whose information we are quering}
}
\value{
Consumed time
}
\description{
This function returns the time consumed so far
}

