% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CResource.R
\name{cResource}
\alias{cResource}
\title{Constructor of CResource objects}
\usage{
cResource(time = NULL, evaluations = NULL, iterations = NULL)
}
\arguments{
\item{time}{Maximum time in seconds. If \code{NULL} (default value), no time limit is set}

\item{evaluations}{Maximum number of evaluations. If \code{NULL} (default value), no evaluations limit is set}

\item{iterations}{Maximum number of iterations. If \code{NULL} (default value), no iterations limit is set}
}
\value{
Object of class \code{\linkS4class{CResource}} with the assigned resources
}
\description{
This function creates objects of class \code{\linkS4class{CResource}}
}

