% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MHResult.R
\name{description}
\alias{description}
\title{Accession function for the \code{description} slot of the object}
\usage{
description(result)
}
\arguments{
\item{result}{Object of class \code{\linkS4class{MHResult}} whose information we are quering}
}
\value{
Content of the \code{description} slot
}
\description{
Accession function for the \code{description} slot of the object
}

