% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EHBSA.R
\docType{class}
\name{EHBSA-class}
\alias{EHBSA-class}
\title{An S4 class to represent distributions based on Edge Histrogam Based Sampling Algorithm.}
\description{
An S4 class to represent distributions based on Edge Histrogam Based Sampling Algorithm.
}
\section{Slots}{

\describe{
\item{\code{adjacencyMatrix}}{A matrix containing the second order marginal probabilities from the data.}
}}

