% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EHBSA.R
\name{ehbsa}
\alias{ehbsa}
\title{Constructor of EHBSA model}
\usage{
ehbsa(data, bratio, ...)
}
\arguments{
\item{data}{The dataframe containing the initial-population to construct the EHBSA.}

\item{...}{Ignored}

\item{bias}{ratio to control the preassure of the learned probabilities. By default it is set to 0.5}
}
\value{
An object of class \code{\linkS4class{EHBSA}} that includes the adjacency matrix of the given data
}
\description{
This function creates an object of class \code{\linkS4class{EHBSA}}
}
\seealso{
Other EDA: \code{\link{bayesianNetwork}},
  \code{\link{generalizedMallowsModel}},
  \code{\link{mallowsModel}}, \code{\link{plackettLuce}},
  \code{\link{univariateMarginals}}
}

