% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/selection.R
\name{elitistSelection}
\alias{elitistSelection}
\title{Elitist selection}
\usage{
elitistSelection(population, evaluation, size = 2, ...)
}
\arguments{
\item{population}{List of individuals where the selection will be applied.}

\item{evaluation}{Vector with the evaluation of the individuals.}

\item{size}{Number of individuals to be selected.}

\item{...}{Ignored}
}
\description{
This function implements the elitist selection, i.e., selection of the best fitted individuals
}
\details{
This function selects the individuals with the lowest evaluation function. If the population contains no individual, \code{NULL} is returned. If the number of individuals is smaller than size, then the whole population is returned.
}
\examples{
url <- system.file("bays29.xml.zip", package="metaheuR")
cost.matrix <- tsplibParser(url)
tsp <- tspProblem(cost.matrix)
n <- ncol(cost.matrix)
rnd.pop <- lapply(1:20,
                  FUN=function(x) {
                    return(randomPermutation(n))
                  })
eval <- unlist(lapply (rnd.pop, FUN=tsp$evaluate))
elitistSelection(population=rnd.pop, evaluation=eval, size=5)
}

