% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Pheromone.R
\name{evaporate}
\alias{evaporate}
\title{Function to uniformly reduce the amount of pheromone}
\usage{
evaporate(object, ...)
}
\arguments{
\item{object}{Object of class \code{\linkS4class{VectorPheromone}} representing the pheromone model to be updated}
}
\description{
This function updates the trail of pheromones through evaporation
}

