% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/01Utils.R
\docType{data}
\name{examplegrid}
\alias{examplegrid}
\title{Example grid}
\format{A GridClass object}
\usage{
examplegrid
}
\description{
An example grid object made of modified Iris-data with package preprocomb.
modifiediris <- droplevels(iris[-c(1:60),])
examplegrid <- preprocomb::setgrid(phases=c("scaling", "smoothing", "outliers", "selection", "sampling"), data=modifiediris)
}
\keyword{datasets}

