% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01Utils.R
\docType{data}
\name{examplehyperparam}
\alias{examplehyperparam}
\title{hyperparameter optimization example}
\format{list}
\usage{
examplehyperparam
}
\description{
An example result to be plotted with plotsearchpath().
}
\details{
examplehyperparam <- metaheurgrid(cores=2, trials=10, iterations=50, nholdout=4) \cr
NOTE: the above uses examplegrid as default grid. Computation time 3,5 hours with \cr
Intel Celeron 1.4 Ghz, 2 cores \cr
plotsearchpath(examplehyperparam) \cr
}
\keyword{datasets}

