% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01Utils.R
\docType{data}
\name{examplesearch}
\alias{examplesearch}
\title{metaheuristic optimization example}
\format{A metaheur class object}
\usage{
examplesearch
}
\description{
An example of metaheuristic search with examplegrid. \cr
50 iterations and 10 times repeated holdout validation.
}
\details{
examplesearch <- metaheur(examplegrid, iterations=50, nholdout=10, cores=2) \cr
}
\keyword{datasets}

