% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExchangeNeighborhood.R
\name{exchangeNeighborhood}
\alias{exchangeNeighborhood}
\title{Basic consturctor of exchange (2-opt) neighborhoods}
\usage{
exchangeNeighborhood(base, random = FALSE)
}
\arguments{
\item{base}{Base solution for the neighborhood. It has to be an object of class \code{\linkS4class{Permutation}}}

\item{random}{A logical value indicating whether the exploration should be done at random}
}
\value{
An object of class \code{\linkS4class{ExchangeNeighobrhood}}
}
\description{
This function creates an object of class \code{\linkS4class{ExchangeNeighborhood}}
}
\seealso{
\code{\link{hasMoreNeighbors}} \code{\link{resetNeighborhood}} \code{\link{nextNeighbor}} \code{\link{swapNeighborhood}}

Other neighborhoods: \code{\link{flipNeighborhood}},
  \code{\link{hammingNeighborhood}},
  \code{\link{hasMoreNeighbors}},
  \code{\link{nextNeighbor}},
  \code{\link{resetNeighborhood}},
  \code{\link{swapNeighborhood}}
}

