% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mutation.R
\name{factor.mutation}
\alias{factor.mutation}
\title{Factor mutation}
\usage{
factor.mutation(solution, ratio, ...)
}
\arguments{
\item{solution}{A factor vector to be mutated.}

\item{ratio}{Ratio of the positions to be mutated. It has to be a value greater than 0.}
}
\description{
This function takes random positions of a factor vector and replaces them with a random value
}
\examples{
A.sol <- factor(rep("A" , 10) , c("A" , "B" , "C"))
factor.mutation (A.sol , ratio = 0.1)
factor.mutation (A.sol , ratio = 0.1)
factor.mutation (A.sol , ratio = 0.5)
}

