% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbor_selectors.R
\name{firstImprovementSelector}
\alias{firstImprovementSelector}
\title{First improvement selection of neighbors}
\usage{
firstImprovementSelector(neighborhood, evaluate, initial.solution,
  initial.evaluation, non.valid = "ignore", valid = function(solution) {    
  TRUE }, correct = function(solution) {     solution },
  resources = cResource(), ...)
}
\arguments{
\item{neighborhood}{Object representing the type of neighborhood to be used}

\item{evaluate}{Function of a single parameter. Given a solution, the function returns its evaluation as a real number}

\item{initial.solution}{Solution whose neighborhood will be explred}

\item{initial.evaluation}{Evaluation value of the initial solution}

\item{non.valid}{Action to be performed when a non valid solution is considered. The options are \code{'ignore'}, meaning that the solution is considered anyway, \code{'discard'}, meaning that the solution is not considered and \code{'correct'}, meaning that the solution has to be corrected. This parameter has to be set only when there can be non valid solutions}

\item{valid}{A function that, given a solution, determines whether it is valid or not}

\item{correct}{A function that, given a non valid solution, corrects it. This optional parameter has to be set only with the \code{'correct'} option}

\item{resources}{Object of class \code{\linkS4class{cResource}} representing the available computational resources for the search}

\item{...}{Special argument to pass additional parameters to the functions used in the search}
}
\value{
The funtion return a list with with three fields, \code{'solution'},  \code{'evaluation'}, contining the best found solution and its evaluation and \code{'resources'}, the remaining computational resources. If the initial solution corresponds to a local optimum, the function returns NULL in the \code{'solution'} field.
}
\description{
This function explores a neighborhood and returns the first solution that improves the current one
}
\details{
In case all the solutions are valid, set the \code{non.valid} parameter to \code{'ignore'} for efficiency. The \code{'neighborhood'} object can be of any type, as long as it has defined
the function \code{'resetNeighborhood'},  \code{'hasMoreNeighbors'} and  \code{'nextNeighbor'}
}
\seealso{
\code{\link{greedySelector}}, \code{\link{basicLocalSearch}}, \code{\link{multistartLocalSearch}},  \code{\link{iteratedLocalSearch}}

Other negibhor selectors: \code{\link{greedySelector}}
}

