% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlipNeighborhood.R
\docType{class}
\name{FlipNeighborhood-class}
\alias{FlipNeighborhood-class}
\title{An S4 class to represent flip neighborhoods for binary vectors}
\description{
An S4 class to represent flip neighborhoods for binary vectors
}
\details{
The new neighbors are generated modifying each position of the vector individually, replacing it with the negation of the current value. The modifications are carried out using the \code{pair.list}, which is iterated using the order defined by \code{index.list}
}
\section{Slots}{

\describe{
\item{\code{base}}{A logical vector representing the base solution whose neighborhood we will explore}

\item{\code{position.list}}{\code{\linkS4class{Permutation}} indicating the oreder in which the exploration is carried out}

\item{\code{random}}{A logical value indicating whether the exploration is at random or not}

\item{\code{id}}{Numeric value indicating the current position (in the \code{position.list}) that will be used to generate a new neighbor}
}}

