% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlipNeighborhood.R
\name{flipNeighborhood}
\alias{flipNeighborhood}
\title{Bais consturctor of flip neighborhoods}
\usage{
flipNeighborhood(base, random = FALSE)
}
\arguments{
\item{base}{Base solution for the neighborhood. It has to be a logical vector}

\item{random}{A logical value indicating whether the exploration should be done at random}
}
\value{
An object of class \code{\linkS4class{FlipNeighobrhood}}
}
\description{
This function creates an object of class \code{\linkS4class{FlipNeighborhood}}
}
\seealso{
\code{\link{hasMoreNeighbors}} \code{\link{resetNeighborhood}} \code{\link{nextNeighbor}}

Other neighborhoods: \code{\link{exchangeNeighborhood}},
  \code{\link{hammingNeighborhood}},
  \code{\link{hasMoreNeighbors}},
  \code{\link{nextNeighbor}},
  \code{\link{resetNeighborhood}},
  \code{\link{swapNeighborhood}}
}

