% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralizedMallowsModel.R
\docType{class}
\name{GeneralizedMallowsModel-class}
\alias{GeneralizedMallowsModel-class}
\title{An S4 class to represent distributions based on the Generalized Mallows Model}
\description{
An S4 class to represent distributions based on the Generalized Mallows Model
}
\section{Slots}{

\describe{
\item{\code{mode}}{The central permutation of the model.}

\item{\code{theta}}{Theta parameters of the model}

\item{\code{distance}}{Type of distance to be used in the model}
}}

