% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cooling_schemes.R
\name{geometric.cooling}
\alias{geometric.cooling}
\title{Geometric decreasing of the temperature}
\usage{
geometric.cooling(initial.temperature, final.temperature, steps)
}
\arguments{
\item{initial.temperature}{Starting temperature}

\item{final.temperature}{Final expected temperature}

\item{steps}{Number of desired steps to get from the initial temperature to the final one}
}
\value{
A function that, given a temperature, returns the updated temperature according to the cooling scheme. The return function uses just a single parameter, the current temperature
}
\description{
This function updates the temperature decreasing it geometrically
}
\examples{
update <- geometric.cooling(initial.temperature = 100 , final.temperature = 10 , steps = 9)
update(100)
update(update(100))
}
\seealso{
Other Cooling schemes: \code{\link{linear.cooling}}
}

