% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MHResult.R
\name{getAlgorithm}
\alias{getAlgorithm}
\title{Accession function for the \code{algorithm} slot of the object}
\usage{
getAlgorithm(result)
}
\arguments{
\item{result}{Object of class \code{\linkS4class{mHResult}} whose information we are quering}
}
\value{
Content of the \code{algorithm} slot
}
\description{
Accession function for the \code{algorithm} slot of the object
}

