% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02Metaheur.R
\name{getbestheur}
\alias{getbestheur}
\title{get the best preprocessing combination}
\usage{
getbestheur(x)
}
\arguments{
\item{x}{(metaheur class object) output of function metaheur()}
}
\value{
(list) best combination, classification accuracy of the best combination
}
\description{
get the best combination and its classification accuracy
}
\examples{
##result <- metaheur(examplegrid)
##getbestheur(result)
}

