% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlackettLuce.R
\name{getColCumsum}
\alias{getColCumsum}
\title{Acumulate the matrix values by columns}
\usage{
getColCumsum(m)
}
\arguments{
\item{m}{matrix}
}
\value{
acumulated matrix
}
\description{
Acumulate the matrix values by columns
}

