% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CResource.R
\name{getConsumedEvaluations}
\alias{getConsumedEvaluations}
\title{Function to get the number of evaluations consumed}
\usage{
getConsumedEvaluations(resource)
}
\arguments{
\item{resource}{Object of class \code{\linkS4class{CResource}} whose information we are quering}
}
\value{
Number of evaluations consumed
}
\description{
This function returns the number of evaluations consumed so far
}

