% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CResource.R
\name{getConsumedTime}
\alias{getConsumedTime}
\title{Function to get the consumed time}
\usage{
getConsumedTime(resource)
}
\arguments{
\item{resource}{Object of class \code{\linkS4class{CResource}} whose information we are quering}
}
\value{
Consumed time
}
\description{
This function returns the time consumed so far
}

