% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MHResult.R
\name{getDescription}
\alias{getDescription}
\title{Accession function for the \code{description} slot of the object}
\usage{
getDescription(result)
}
\arguments{
\item{result}{Object of class \code{\linkS4class{mHResult}} whose information we are quering}
}
\value{
Content of the \code{description} slot
}
\description{
Accession function for the \code{description} slot of the object
}

