% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHResult.R
\name{getEvaluatedSolutions}
\alias{getEvaluatedSolutions}
\title{Accession function for the \code{solutions} slot of the object}
\usage{
getEvaluatedSolutions(result)
}
\arguments{
\item{result}{Object of class \code{\linkS4class{mHResult}} whose information we are quering}
}
\value{
Content of the \code{solutions} slot
}
\description{
Accession function for the \code{solutions} slot of the object
}

