% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MHResult.R
\name{getEvaluation}
\alias{getEvaluation}
\title{Accession function for the \code{evaluation} slot of the object}
\usage{
getEvaluation(result)
}
\arguments{
\item{result}{Object of class \code{\linkS4class{mHResult}} whose information we are quering}
}
\value{
Content of the \code{evaluation} slot
}
\description{
Accession function for the \code{evaluation} slot of the object
}

