% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02Metaheur.R
\name{getlogs}
\alias{getlogs}
\title{get the logs of a metaheur search}
\usage{
getlogs(x, n = NULL)
}
\arguments{
\item{x}{(metaheur class object) output of function metaheur()}

\item{n}{(integer) number of log entries show, default to all entries}
}
\description{
get the logs of a metaheur search
}
\examples{
##getlogs(examplesearch)
}

