% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHResult.R
\name{getProgress}
\alias{getProgress}
\title{Accession function for the \code{progress} slot of the object}
\usage{
getProgress(result)
}
\arguments{
\item{result}{Object of class \code{\linkS4class{mHResult}} whose information we are quering}
}
\value{
Content of the \code{progress} slot
}
\description{
Accession function for the \code{progress} slot of the object
}

