% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CResource.R
\name{getRemainingEvaluations}
\alias{getRemainingEvaluations}
\title{Function to get the remaining evaluations available}
\usage{
getRemainingEvaluations(resource)
}
\arguments{
\item{resource}{Object of class \code{\linkS4class{CResource}} whose information we are quering}
}
\value{
Remaining number of evaluations
}
\description{
This function returns the remaining evaluations available
}

