% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CResource.R
\name{getRemainingTime}
\alias{getRemainingTime}
\title{Function to get the remaining time available}
\usage{
getRemainingTime(resource)
}
\arguments{
\item{resource}{Object of class \code{\linkS4class{CResource}} whose information we are quering}
}
\value{
Remaining time
}
\description{
This function returns the remaining time available
}

