% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MHResult.R
\name{getResources}
\alias{getResources}
\title{Accession function for the \code{resources} slot of the object}
\usage{
getResources(result)
}
\arguments{
\item{result}{Object of class \code{\linkS4class{mHResult}} whose information we are quering}
}
\value{
Content of the \code{resources} slot
}
\description{
Accession function for the \code{resources} slot of the object
}

