% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02Metaheur.R
\name{getwalltime}
\alias{getwalltime}
\title{get the execution wall-clock time}
\usage{
getwalltime(x)
}
\arguments{
\item{x}{(metaheur class object) output of function metaheur()}
}
\description{
get the execution wall-clock time
}
\examples{
##getwalltime(examplesearch)
}

