% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExchangeNeighborhood.R, R/HammingNeighborhood.R
\name{hammingNeighborhood}
\alias{hammingNeighborhood}
\title{Basic consturctor of exchange (2-opt) neighborhoods}
\usage{
hammingNeighborhood(base, random = FALSE)

hammingNeighborhood(base, random = FALSE)
}
\arguments{
\item{base}{Base solution for the neighborhood. It has to be an object of class \code{\linkS4class{Permutation}}}

\item{random}{A logical value indicating whether the exploration should be done at random}

\item{base}{Base solution for the neighborhood. It has to be a vector of factors (all with the same levels)}

\item{random}{A logical value indicating whether the exploration should be done at random}
}
\value{
An object of class \code{\linkS4class{ExchangeNeighobrhood}}

An object of class \code{\linkS4class{HammingNeighobrhood}}
}
\description{
This function creates an object of class \code{\linkS4class{ExchangeNeighborhood}}

This function creates an object of class \code{\linkS4class{HammingNeighborhood}}
}
\seealso{
\code{\link{hasMoreNeighbors}} \code{\link{resetNeighborhood}} \code{\link{nextNeighbor}} \code{\link{swapNeighborhood}}

\code{\link{hasMoreNeighbors}} \code{\link{resetNeighborhood}} \code{\link{nextNeighbor}}

Other neighborhoods: \code{\link{flipNeighborhood}};
  \code{\link{has.more.neighbors}};
  \code{\link{next.neighbor}};
  \code{\link{reset.neighborhood}};
  \code{\link{swapNeighborhood}}

Other neighborhoods: \code{\link{flipNeighborhood}};
  \code{\link{has.more.neighbors}};
  \code{\link{next.neighbor}};
  \code{\link{reset.neighborhood}};
  \code{\link{swapNeighborhood}}
}

