% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Neighborhood.R
\name{hasMoreNeighbors}
\alias{hasMoreNeighbors}
\title{Function to check whether we have accessed to all the solutions in the neighborhood or not}
\usage{
hasMoreNeighbors(neighborhood)
}
\arguments{
\item{neighborhood}{Object representing the neighborhood}
}
\value{
\code{TRUE} if there are more solutions to explore, \code{FALSE} in other case
}
\description{
This is one of the functions that has to be implemented by any object to be used as a neighborhood
}
\seealso{
\code{\link{nextNeighbor}} \code{\link{resetNeighborhood}}

Other neighborhoods: \code{\link{exchangeNeighborhood}},
  \code{\link{flipNeighborhood}},
  \code{\link{hammingNeighborhood}},
  \code{\link{nextNeighbor}},
  \code{\link{resetNeighborhood}},
  \code{\link{swapNeighborhood}}
}

