% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CResource.R
\name{isFinished}
\alias{isFinished}
\title{Function check whether the computational resources are finished}
\usage{
isFinished(resource)
}
\arguments{
\item{resource}{Object of class \code{\linkS4class{CResource}} to which the values will be substracted}
}
\value{
A logical value, \code{TRUE} if any of the resources is finished
}
\description{
This function determines what (time, evaluations and/or iterations) is finished
}

