% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/crossover.R
\name{k.point.crossover}
\alias{k.point.crossover}
\title{k point crossover}
\usage{
k.point.crossover(sol.1, sol.2, k = 1, ...)
}
\arguments{
\item{sol.1}{First vector to be crossed}

\item{sol.2}{Second vector to be crossed}

\item{k}{Number of cut points. It has to be a number between 1 and the size of the vectors minus 1}

\item{...}{Ignored}
}
\description{
This function implements a k point crossover for vectors
}
\details{
The vectors are splited into k+1 fragments and new solutions are constructed alternating elements from each vector
}
\examples{
A.sol <- rep("A" , 10)
B.sol <- rep("B" , 10)
k.point.crossover (sol.1 = A.sol , sol.2 = B.sol , k = 1)
k.point.crossover (sol.1 = A.sol , sol.2 = B.sol , k = 2)
k.point.crossover (sol.1 = A.sol , sol.2 = B.sol , k = 9)
}

