% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_problems.R
\name{knapsackProblem}
\alias{knapsackProblem}
\title{Knapsac problem evaluator}
\usage{
knapsackProblem(weight, value, limit)
}
\arguments{
\item{weight}{Vector with the weights associated to each item}

\item{value}{Vector with the value associated to each item}

\item{limit}{Maximum weight allowed}
}
\value{
A list of functions to be used to solve a knapsack problem. This includes the functions \code{evaluate}, for the evaluation of a solution, \code{valid}, to check whetehr a solution is valid or not and 'correct', to correct a non-valid solution; all the functions have a single argument,  \code{solution}, representing the solution considered. Note that, given that the goal in all the algorithms in the library is minimizing the objective function, the \code{evaluate}. The solutions have to be vectors of logic values, indicating which items are included in the solution.
}
\description{
This function generates an evaluation, validity and correction functions associated  with a classical knapsack problem
}
\examples{
## Problem of size 100 with random vectors and limit
n <- 100
w <- runif(n)
v <- runif(n)
l <- sum(w[runif(n) > 0.5])
knp <- knapsackProblem(w, v, l)
# Use of the functioins
rnd.solution <- runif(n) > 0.5
knp$evaluate(rnd.solution)
knp$valid(rnd.solution)
corrected.solution <- knp$correct (rnd.solution)
knp$valid(corrected.solution)
}
\seealso{
Other Problems: \code{\link{graphColoringProblem}},
  \code{\link{lopProblem}}, \code{\link{mdsProblem}},
  \code{\link{misProblem}}, \code{\link{pfspProblem}},
  \code{\link{qapProblem}}, \code{\link{tspProblem}}
}

