% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cooling_schemes.R
\name{linear.cooling}
\alias{linear.cooling}
\title{Linear decreasing of the temperature}
\usage{
linear.cooling(initial.temperature, final.temperature, steps)
}
\arguments{
\item{initial.temperature}{Starting temperature}

\item{final.temperature}{Final expected temperature}

\item{steps}{Number of desired steps to get from the initial temperature to the final one}
}
\value{
A function that, given a temperature, returns the updated temperature according to the cooling scheme. The return function uses just a single parameter, the current temperature
}
\description{
This function updates the temperature decreasing it linearly
}
\examples{
boltzmann.accept (delta = -1 , temperature = 2)
boltzmann.accept (delta = -3 , th = -2)
boltzmann.accept (delta = -1 , th = 0)
}

