% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MallowsModel.R
\docType{class}
\name{MallowsModel-class}
\alias{MallowsModel-class}
\title{An S4 class to represent distributions based the Mallows Model}
\description{
An S4 class to represent distributions based the Mallows Model
}
\section{Slots}{

\describe{
\item{\code{mode}}{The central permutation of the model.}

\item{\code{theta}}{Theta parameter of the model}

\item{\code{distance}}{Type of distance to be used in the model}
}}

