% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MallowsModel.R
\name{mallowsModel}
\alias{mallowsModel}
\title{Baisic consturctor of Mallows models}
\usage{
mallowsModel(data, distance = "Kendall", tmin = NULL, tmax = NULL, ...)
}
\arguments{
\item{data}{A list with the permutations from where the model will be learned}

\item{distance}{Type of distance to be used in the model. By default, it is set to \code{Kendall}}

\item{tmin}{Minimum value for the theta parameter. If \code{NULL} (default value) no bound is set}

\item{tmax}{Maximum value for the theta parameter. If \code{NULL} (default value) no bound is set}

\item{...}{Ignored}
}
\value{
An object of class \code{\linkS4class{MallowsModel}} that represents the learned model
}
\description{
This function creates an object of class \code{\linkS4class{MallowsModel}}
}
\seealso{
Other EDA: \code{\link{bayesianNetwork}},
  \code{\link{ehbsa}},
  \code{\link{generalizedMallowsModel}},
  \code{\link{plackettLuce}},
  \code{\link{univariateMarginals}}
}

