% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graph_problems.R
\name{mdsProblem}
\alias{mdsProblem}
\title{Minimum dominating set}
\usage{
mdsProblem(graph, penalization = 0)
}
\arguments{
\item{graph}{Graph where we have to find the minimum dominating set (MDS)}
}
\value{
A list of functions to be used to solve a MDS problem. This includes the functions \code{evaluate}, for the evaluation of a solution, \code{valid}, to check whetehr a solution is valid or not, \code{correct}, to correct a non-valid solution and \code{plot} to graphically show the solution; all the functions have a single argument,  \code{solution}. The solutions passed to these functions has to be a logical vector indicating with \code{TRUE} which nodes are in the independent set.
}
\description{
This function generates an evaluation, validity and correction functions associated  with a classical minimum dominating problem
}
\details{
The evaluation function includes another parameter, \code{penalization}, which can be used to penalize non-valid solutions. The penalization terms is the number of nodes that are not connected to the nodes in the solution, and it is weighted with the factor passed in the \code{penalization} parameter. By default its value is 0.
}
\examples{
library("igraph")
}

