% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02Metaheur.R
\docType{class}
\name{metaheur-class}
\alias{metaheur-class}
\title{An S4 class to represent metaheuristic optimization results}
\description{
An S4 class to represent metaheuristic optimization results
}
\section{Slots}{

\describe{
\item{\code{solutionlist}}{(list) of best solutions, i.e. preprocessing combinations for each start}

\item{\code{iterationhistory}}{(list) of iteration classification accuracies for each start}

\item{\code{logs}}{(list) event log for each start}

\item{\code{walltime}}{(integer) execution wall-clock time in minutes}
}}

