\name{metaheuR-package}
\alias{metaheuR-package}
\alias{metaheuR}
\docType{package}
\title{
metaheuR: Metaheuristics for teaching in R
}
\description{
This package has been designed to be used for teaching metaheuristics. Therefore, the main concern is the readability of the code, rather than the performance
}
\details{
\tabular{ll}{
Package: \tab metaheuR\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2011-01-29\cr
License: \tab GPL\cr
}
This package has been designed to be used for teaching metaheuristics. Therefore, the main concern is the readability of the code, rather than the performance
}
\author{
Borja Calvo <borja.calvo@ehu.eus>
Maintainer: Borja Calvo <borja.calvo@ehu.eus>
}
