% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/02Metaheur.R
\name{metaheurcompare}
\alias{metaheurcompare}
\title{metaheurcompare}
\usage{
metaheurcompare(gridclassobject = examplegrid, runs = 2, iterations = 10,
  initialtemperature = c(0.01, 1), tempconst = c(1, 0.85),
  reheat = c(0.01, 0.1), taboolistlength = c(1, 1), late = c(0, 1))
}
\arguments{
\item{gridclassobject}{(GridClass) created by setgrid function in preprocomb package, defaults to examplegrid}

\item{runs}{(integer) number of runs, defaults to two}

\item{iterations}{(integer) number of iterations done for a restart}

\item{initialtemperature}{(numeric vector) initial propability for acccepting an inferior candidate, between 0 and 1}

\item{tempconst}{(numeric vector) multiplier for decreasing temperature on each iteration}

\item{reheat}{(numeric vector) propability of increasing temperature on each iteration}

\item{taboolistlength}{(integer vector) number of previous solution that can not be revisited, > 1}

\item{late}{(integer vector) location of previous best solution a candidate is compared to, default to 0 for last}
}
\description{
Compare parameter settings
}
\examples{
## result <- metaheur(examplegrid,verbose=TRUE)
}

