% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03MetaheurHyperparameters.R
\docType{class}
\name{metaheurhyper-class}
\alias{metaheurhyper-class}
\title{An S4 class to represent hyperparameter optimization results}
\description{
An S4 class to represent hyperparameter optimization results
}
\section{Slots}{

\describe{
\item{\code{res}}{(matrix) of classification accuracy results, rows for iterations and columns for parameter combinations}

\item{\code{namegrid}}{(list) parameter combinations}

\item{\code{baseheur}}{(list) metaheur objects}

\item{\code{walltime}}{(integer) execution wall-clock time in minutes}
}}

