% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03MetaheurHyperparameters.R
\name{metaheurhyper}
\alias{metaheurhyper}
\title{search for best hyperparameters for metaheuristic optimization}
\usage{
metaheurhyper(gridclassobject = examplegrid, searchtype = "grid",
  nrohyperparams = 3, iterations = 10, cores = 1, nholdout = 2,
  trials = 3, model = "rpart")
}
\arguments{
\item{gridclassobject}{(GridClass) created by setgrid function in preprocomb package, defaults to examplegrid}

\item{searchtype}{(character) grid or random search}

\item{nrohyperparams}{(integer) number of hyperparameters used in random search, between 1 and 5, default to 3}

\item{iterations}{(integer) number of iterations done for a restart}

\item{cores}{(integer) number of cores used in computation of classification accuracies holdout rounds}

\item{nholdout}{(integer) number of holdout rounds in computing classification accuracies}

\item{trials}{(integer) number of trials}

\item{model}{(character) caret name of predictive model, defaults to "rpart"}
}
\description{
search for best hyperparameters for metaheuristic optimization
}
\examples{
## result <- metaheurhyper(cores=2, trials=2, iterations=30)
}

