% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHResult.R
\docType{class}
\name{MHResult-class}
\alias{MHResult-class}
\title{An S4 class to represent the output of a search heuristic}
\description{
An S4 class to represent the output of a search heuristic
}
\section{Slots}{

\describe{
\item{\code{description}}{A free filed to include comments}

\item{\code{algorithm}}{Name of the algorithm used}

\item{\code{parameters}}{This should be a named list with the parameters used by the algorithm}

\item{\code{solution}}{Best solution found}

\item{\code{evaluation}}{Evaluation of the best solution(s)}

\item{\code{resources}}{Object contining the information about the available and consumed computational resources}

\item{\code{log}}{A numeric matrix containing the evolution of the search. The matrix should have four columns, named as \code{Time}, \code{Evaluations}, \code{Iterations}, \code{Current_sol}, \code{Current_sd} and \code{Best_sol}}

\item{\code{solutions}}{A list with all the solutions evaluated during the search. Each element of the list is a list with two fields, 'iteration', 'solution' and 'evaluation'. The former contains the iteration when the soulution has been evaluated, the second the solution itself and the last one its evaluation.}
}}

