% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MHResult.R
\name{mhresult}
\alias{mhresult}
\title{Function to construct objects of type \code{MHResult}}
\usage{
mhresult(algorithm, description = "", parameters, optima = NULL, evaluation,
  resources, log = matrix())
}
\arguments{
\item{algorithm}{String indicating the algorithm run}

\item{description}{Description of the algorithm run}

\item{parameters}{List of parameters used in the experiment}

\item{optima}{One or more optimal solutions}

\item{evaluation}{Optimal value of the objective function}

\item{resources}{Object of class \code{\link{CResource}} contining the information about the use of resources in the experiment}

\item{log}{Matrix with the evolution of the search}
}
\value{
A new object of class \code{\link{MHResult}} with the updated remaining resources
}
\description{
This function creates objects of class \code{\link{MHResult}}
}

