% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Neighborhood.R
\name{next.neighbor}
\alias{next.neighbor}
\title{Function to obtain a new neighbor in a neighborhood}
\usage{
next.neighbor(neighborhood)
}
\arguments{
\item{neighborhood}{Object representing the neighborhood}
}
\value{
A new solution in th neighborhood represented by \code{neighborhood}
}
\description{
This is one of the functions that has to be implemented by any object to be used as a neighborhood
}
\seealso{
\code{\link{hasMoreNeighbors}} \code{\link{resetNeighborhood}}

Other neighborhoods: \code{\link{flipNeighborhood}};
  \code{\link{hammingNeighborhood}},
  \code{\link{hammingNeighborhood}};
  \code{\link{has.more.neighbors}};
  \code{\link{reset.neighborhood}};
  \code{\link{swapNeighborhood}}
}

