% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Neighborhood.R
\name{nextNeighbor}
\alias{nextNeighbor}
\title{Function to obtain a new neighbor in a neighborhood}
\usage{
nextNeighbor(neighborhood)
}
\arguments{
\item{neighborhood}{Object representing the neighborhood}
}
\value{
A new solution in th neighborhood represented by \code{neighborhood}
}
\description{
This is one of the functions that has to be implemented by any object to be used as a neighborhood
}
\seealso{
\code{\link{hasMoreNeighbors}} \code{\link{resetNeighborhood}}

Other neighborhoods: \code{\link{exchangeNeighborhood}};
  \code{\link{flipNeighborhood}};
  \code{\link{hammingNeighborhood}};
  \code{\link{hasMoreNeighbors}};
  \code{\link{resetNeighborhood}};
  \code{\link{swapNeighborhood}}
}

