% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MHResult.R
\name{optima}
\alias{optima}
\title{Accession function for the \code{optima} slot of the object}
\usage{
optima(result)
}
\arguments{
\item{result}{Object of class \code{\linkS4class{MHResult}} whose information we are quering}
}
\value{
Content of the \code{optima} slot
}
\description{
Accession function for the \code{optima} slot of the object
}

