% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/crossover.R
\name{order.crossover}
\alias{order.crossover}
\title{Order crossover for permutations}
\usage{
order.crossover(sol.1, sol.2, ...)
}
\arguments{
\item{sol.1}{First vector to be crossed}

\item{sol.2}{Second vector to be crossed}

\item{...}{Ignored}
}
\description{
This function implements the order crossover operator for permutations
}
\details{
Two random points are selected and the values between this points are directly copied from one of the parents. The rest of the positions are filled using the order in the second parent
}
\examples{
A.sol <- identity.permutation (10)
B.sol <- random.permutation (10)
order.crossover (sol.1 = A.sol , sol.2 = B.sol)
}

