% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MHResult.R
\name{parameters}
\alias{parameters}
\title{Accession function for the \code{parameters} slot of the object}
\usage{
parameters(result)
}
\arguments{
\item{result}{Object of class \code{\linkS4class{MHResult}} whose information we are quering}
}
\value{
Content of the \code{parameters} slot
}
\description{
Accession function for the \code{parameters} slot of the object
}

