% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermuLinkPheromone.R
\docType{class}
\name{PermuLinkPheromone-class}
\alias{PermuLinkPheromone-class}
\title{An S4 class to basic link-based pheromone model for permutations}
\description{
An S4 class to basic link-based pheromone model for permutations
}
\section{Slots}{

\describe{
\item{\code{trail}}{A matrix with the pheromone values for each link. Only the upper triangle of the matrix is considered}

\item{\code{evaporation.factor}}{A value between 0 and 1 (both excluded) that indicates the factor used in the evaporation phase}
}}

