% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PermuLinkPheromone.R
\name{permuLinkPheromone}
\alias{permuLinkPheromone}
\title{Bais consturctor of pheromone models based on the view of permutations as Hamiltonian cycles in complete graphs}
\usage{
permuLinkPheromone(initial.trail, evaporation.factor, ...)
}
\arguments{
\item{initial.trail}{An square matrix containing the intial pheromone value for each link in the complete graph}

\item{evaporation.factor}{A number between 0 and 1 (both excluded) that represents the factor used in the evaporation phase}

\item{...}{Ignored}
}
\value{
An object of class \code{\linkS4class{PermuLinkPheromone}}
}
\description{
This function creates an object of class \code{\linkS4class{PermuLinkPheromone}}
}
\seealso{
Other ACO: \code{\link{permuPosPheromone}};
  \code{\link{vectorPheromone}}
}

