% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermuPosPheromone.R
\docType{class}
\name{PermuPosPheromone-class}
\alias{PermuPosPheromone-class}
\title{An S4 class to basic position-based pheromone model for permutations}
\description{
An S4 class to basic position-based pheromone model for permutations
}
\section{Slots}{

\describe{
\item{\code{trail}}{A matrix with the pheromone values for each position. The columns represent the position and the rows the value.}

\item{\code{evaporation.factor}}{A value between 0 and 1 (both excluded) that indicates the factor used in the evaporation phase}
}}

