% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PermuPosPheromone.R
\name{permuPosPheromone}
\alias{permuPosPheromone}
\title{Bais consturctor of pheromone models for permutation based on position}
\usage{
permuPosPheromone(initial.trail, evaporation.factor, ...)
}
\arguments{
\item{initial.trail}{An square matrix containing the intial pheromone value for combination of position and value}

\item{evaporation.factor}{A number between 0 and 1 (both excluded) that represents the factor used in the evaporation phase}

\item{...}{Ignored}
}
\value{
An object of class \code{\linkS4class{PermuPosPheromone}}
}
\description{
This function creates an object of class \code{\linkS4class{PermuPosPheromone}}
}
\seealso{
Other ACO: \code{\link{permuLinkPheromone}};
  \code{\link{vectorPheromone}}
}

