% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Permutation.R
\docType{class}
\name{Permutation-class}
\alias{Permutation-class}
\title{An S4 class to represent permutations.}
\description{
An S4 class to represent permutations.
}
\section{Slots}{

\describe{
\item{\code{permutation}}{A numeric vector containing the permutation. This vector has to be a valid premutation}
}}

