% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{pfspParser}
\alias{pfspParser}
\title{Function to read PFSP files}
\usage{
pfspParser(file)
}
\arguments{
\item{file}{File containing the information}
}
\value{
A matrix with the job length in each machine
}
\description{
This function reads plain text files containing PFSP
}
\seealso{
\code{\link{pfspProblem}}

Other parsers: \code{\link{tsplibParser}}
}

