% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_problems.R
\name{pfspProblem}
\alias{pfspProblem}
\title{PFSP problem evaluator}
\usage{
pfspProblem(job.machine.matrix)
}
\arguments{
\item{job.machine.matrix}{Matrix containing the time required by each job in each machine. 
Each row is a machine and each column a job}
}
\value{
A list with three elements,  \code{evaluate}, a function to evaluate
solutions, \code{num.jobs}, the number of jobs, \code{num.machines}, the number of machines
and, \code{plotSolution}, a function to plot a solution for the problem.
}
\description{
This function generates an evaluation function associated with a PFSP problem
}
\examples{
cmatrix <- matrix(runif(100), ncol=5)
tsp <- pfspProblem(cmatrix)
tsp$evaluate(randomPermutation(10))

coord <- matrix(runif(14), ncol=2)
rownames(coord) <- paste0("C", 1:7)
cmat <- as.matrix(dist(coord))
tsp <- tspProblem(cmat, coord)
tsp$plotSolution(randomPermutation(7), plot.names=TRUE)

}
\seealso{
Other Problems: \code{\link{graphColoringProblem}},
  \code{\link{knapsackProblem}}, \code{\link{lopProblem}},
  \code{\link{mdsProblem}}, \code{\link{misProblem}},
  \code{\link{qapProblem}}, \code{\link{tspProblem}}
}

