% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlackettLuce.R
\docType{class}
\name{PlackettLuce-class}
\alias{PlackettLuce-class}
\title{An S4 class to represent distributions based on Plackett Luce distribution model.}
\description{
An S4 class to represent distributions based on Plackett Luce distribution model.
}
\section{Slots}{

\describe{
\item{\code{parameters}}{The weight parameters of the learned population.}
}}

