% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlackettLuce.R
\name{plackettLuce}
\alias{plackettLuce}
\title{This function creates an object of class \code{\linkS4class{PlackettLuce}}}
\usage{
plackettLuce(data, maxIter, ...)
}
\arguments{
\item{data}{The dataframe containing the initial-population to contruct the Plackett-Luce model.}

\item{maxIter}{Optional. Maximum iterations for the MM algorithm.}

\item{...}{Ignored}
}
\value{
An object of class \code{\linkS4class{PlackettLuce}} that includes the weight parameter vector of the given data
}
\description{
This function creates an object of class \code{\linkS4class{PlackettLuce}}
}
\seealso{
Other EDA: \code{\link{bayesianNetwork}},
  \code{\link{ehbsa}}, \code{\link{mallowsModel}},
  \code{\link{univariateMarginals}}
}

