% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MHResult.R
\name{plot.correlation}
\alias{plot.correlation}
\title{Function to plot the correlation among evaluations, iterations and time}
\usage{
plot.correlation(result, a = "time", b = "evaluations", ...)
}
\arguments{
\item{result}{Object of class \code{\link{MHResult}} whose log will be visualized}

\item{a}{Parameter indicating the value for the X axis. Valid options are 'time', 'evaluations' and 'iterations'}

\item{b}{Parameter indicating the value for the Y axis. Valid options are 'time', 'evaluations' and 'iterations'}

\item{...}{Additional parameters for the function \code{\link{geom_point}} used to draw the plots}
}
\value{
An object of class \code{\link{ggplot}}
}
\description{
This function produces an object of class \code{\link{ggplot}} showing the correlation among evaluations, iterations and time
}

