% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MHResult.R
\name{plot.progress}
\alias{plot.progress}
\title{Function to plot the evolution of the evaluation function during the search}
\usage{
plot.progress(result, vs = "evaluations", ...)
}
\arguments{
\item{result}{Object of class \code{\link{MHResult}} whose log will be visualized}

\item{vs}{Parameter indicating the value for the X axis. Valid options are 'time', 'evaluations' and 'iterations'}

\item{...}{Additional parameters for the function \code{\link{geom_line}} used to draw the line}
}
\value{
An object of class \code{\link{ggplot}}
}
\description{
This function produces an object of class \code{\link{ggplot}} with the evolution of the search
}

