% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/03MetaheurCompare.R
\name{plotdensity}
\alias{plotdensity}
\title{plotdensity}
\usage{
plotdensity(x)
}
\arguments{
\item{x}{output of metaheurcompare function}
}
\description{
plot the density distribution of classification accuracies of metaheurcompare()
}
\examples{
## result <- metaheurcompare()
## plotdensity(result)
}

