% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MHResult.R
\name{plotProgress}
\alias{plotProgress}
\title{Function to plot the evolution of the evaluation function during the search}
\usage{
plotProgress(result, x = "evaluations", y = "current", type = "line", ...)
}
\arguments{
\item{result}{Object of class \code{\linkS4class{mHResult}} whose log will be visualized or a named list of objects}

\item{x}{Parameter indicating the value for the X axis. Valid options are 'time', 'evaluations' and 'iterations'}

\item{y}{Parameter indicating the value for the Y axis. Valid options are 'current' and 'best'}

\item{...}{Additional parameters for the function \code{\link{geom_line}} used to draw the line}
}
\value{
An object of class \code{\link{ggplot}}
}
\description{
This function produces an object of class \code{\link{ggplot}} with the evolution of the search
}

