% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06ExtractComponent.R
\name{plotsearchpath}
\alias{plotsearchpath}
\title{plot search path}
\usage{
plotsearchpath(x)
}
\arguments{
\item{x}{(metaheur object or list of metaheurgrid class objects) object to be plotted}
}
\description{
Plots the search path, i.e. classification accuracy by iteration.
metaheur class plots (output of metaheur function) can include restarts.
List of metaheurhyper class objects (output of metaheurhyper function) plot
includes best, median and worst scenarious.
}
\examples{
plotsearchpath(examplesearch)
plotsearchpath(examplehyperparam)
}

