% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MHResult.R
\name{progress}
\alias{progress}
\title{Accession function for the \code{progress} slot of the object}
\usage{
progress(result)
}
\arguments{
\item{result}{Object of class \code{\link{MHResult}} whose information we are quering}
}
\value{
Content of the \code{progress} slot
}
\description{
Accession function for the \code{progress} slot of the object
}

