% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/permutation_problems.R
\name{qap.problem}
\alias{qap.problem}
\title{QAP problem evaluator}
\usage{
qap.problem(fmatrix, dmatrix)
}
\arguments{
\item{fmatrix}{Flow matrix for the QAP problem}

\item{dmatrix}{Distance matrix for the QAP problem}
}
\value{
A function that can be used to evaluate solutions for a QAP problem
}
\description{
This function generates an evaluation function associated
with a QAP problem
}
\examples{
fmatrix<-matrix(runif(100),ncol=10)
dmatrix<-matrix(runif(100),ncol=10)
qap<-qap.problem(fmatrix, dmatrix)
qap$evaluate(random.permutation(10))
}

