% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_problems.R
\name{qapProblem}
\alias{qapProblem}
\title{QAP problem evaluator}
\usage{
qapProblem(fmatrix, dmatrix)
}
\arguments{
\item{fmatrix}{Flow matrix for the QAP problem}

\item{dmatrix}{Distance matrix for the QAP problem}
}
\value{
A function that can be used to evaluate solutions for a QAP problem
}
\description{
This function generates an evaluation function associated
with a QAP problem
}
\examples{
fmatrix <- matrix(runif(100), ncol=10)
dmatrix <- matrix(runif(100), ncol=10)
qap <- qapProblem(fmatrix, dmatrix)
qap$evaluate(randomPermutation(10))

}

