% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CResource.R
\name{remaining.time}
\alias{remaining.time}
\title{Function to get the remaining time available}
\usage{
remaining.time(resource)
}
\arguments{
\item{resource}{Object of class \code{\link{CResource}} whose information we are quering}
}
\value{
Remaining time
}
\description{
This function returns the remaining time available
}

