% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Neighborhood.R
\name{reset.neighborhood}
\alias{reset.neighborhood}
\title{Function to reset a neighborhood to a given base solution}
\usage{
reset.neighborhood(neighborhood, solution)
}
\arguments{
\item{neighborhood}{Object representing the neighborhood}

\item{solution}{Solution whose neighborhood we want to explore}
}
\description{
This is one of the functions that has to be implemented by any object to be used as a neighborhood
}
\seealso{
\code{\link{nextNeighbor}} \code{\link{resetNeighborhood}}

Other neighborhoods: \code{\link{has.more.neighbors}};
  \code{\link{next.neighbor}}
}

