% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Neighborhood.R
\name{resetNeighborhood}
\alias{resetNeighborhood}
\title{Function to reset a neighborhood to a given base solution}
\usage{
resetNeighborhood(neighborhood, solution)
}
\arguments{
\item{neighborhood}{Object representing the neighborhood}

\item{solution}{Solution whose neighborhood we want to explore}
}
\description{
This is one of the functions that has to be implemented by any object to be used as a neighborhood
}
\seealso{
\code{\link{nextNeighbor}} \code{\link{resetNeighborhood}}

Other neighborhoods: \code{\link{exchangeNeighborhood}},
  \code{\link{flipNeighborhood}},
  \code{\link{hammingNeighborhood}},
  \code{\link{hasMoreNeighbors}},
  \code{\link{nextNeighbor}},
  \code{\link{swapNeighborhood}}
}

