% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/continuous_problems.R
\name{rosenbrock.problem}
\alias{rosenbrock.problem}
\title{Rosenbrock function to evaluate numeric algorithms}
\usage{
rosenbrock.problem(size)
}
\arguments{
\item{size}{Number of dimensions}
}
\value{
A function that can be used to evaluate solutions
}
\description{
This function implements Rosenbrock's function
}
\examples{
rs<-rosenbrock.problem(3)
rs$evaluate(c(2,-1,1))
rs$evaluate(c(2,1,1))
# The only optimum for three dimensions is at 1,1,1
rs$evaluate(c(1,1,1))
}

