% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_problems.R
\name{rosenbrockProblem}
\alias{rosenbrockProblem}
\title{Rosenbrock function to evaluate numeric algorithms}
\usage{
rosenbrockProblem(size)
}
\arguments{
\item{size}{Number of dimensions}
}
\value{
A list with two elements, the evaluation function (\code{evaluate}) and the number of dimensions (\code{size})
}
\description{
This function implements Rosenbrock's function
}
\examples{
rs <- rosenbrockProblem(3)
rs$evaluate(c(2, -1, 1))
rs$evaluate(c(2, 1, 1))
# The only optimum for three dimensions is at 1,1,1
rs$evaluate(c(1, 1, 1))

}

