% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/selection.R
\name{roulette.selection}
\alias{roulette.selection}
\title{Roulette selection}
\usage{
roulette.selection(population, evaluation, size = 2, use.rankings, ...)
}
\arguments{
\item{population}{List of individuals where the selection will be applied.}

\item{evaluation}{Vector with the evaluation of the individuals.}

\item{size}{Number of individuals to be selected.}

\item{...}{Ignored}
}
\description{
This function implements the selection based on a biased roulette
}
\details{
This function selects the individuals with the lowest evaluation function. If the population contains no individual, \code{NULL} is returned.
}
\examples{
url <- system.file("bays29.xml.zip", package = "metaheuR")
cost.matrix <- tsplib.parser(url)
tsp <- tsp.problem(cost.matrix)
n <- ncol(cost.matrix)
rnd.pop <- lapply(1:20 , FUN = function(x) random.permutation(n))
eval <- unlist(lapply (rnd.pop , FUN = tsp$evaluate))
roulette.selection(population = rnd.pop , evaluation = eval , size = 5 , use.rankings = TRUE)
roulette.selection(population = rnd.pop , evaluation = eval , size = 9 , use.rankings = FALSE)
}

