% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{rouletteSelection}
\alias{rouletteSelection}
\title{Roulette selection}
\usage{
rouletteSelection(population, evaluation, use.rankings, size = 2, ...)
}
\arguments{
\item{population}{List of individuals where the selection will be applied.}

\item{evaluation}{Vector with the evaluation of the individuals.}

\item{size}{Number of individuals to be selected.}

\item{...}{Ignored}
}
\description{
This function implements the selection based on a biased roulette
}
\details{
This function selects the individuals with the lowest evaluation function. If the population contains no individual, \code{NULL} is returned.
}
\examples{

url <- system.file("bays29.xml.zip", package="metaheuR")
cost.matrix <- tsplibParser(url)
tsp <- tspProblem(cost.matrix)
n <- ncol(cost.matrix)
rnd.pop <- lapply(1:20, 
                  FUN=function(x) {
                    return(randomPermutation(n))
                  })
eval <- unlist(lapply(rnd.pop, FUN=tsp$evaluate))
rouletteSelection(population=rnd.pop, evaluation=eval, size=5, use.rankings=TRUE)
rouletteSelection(population=rnd.pop, evaluation=eval, size=9, use.rankings=FALSE)

}

