% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ls_acceptance.R
\name{simulated.annealing}
\alias{simulated.annealing}
\title{Simulated annealing}
\usage{
simulated.annealing(evaluate, initial.solution, neighborhood, cooling.scheme,
  initial.temperature, final.temperature, eq.criterion = "evaluations",
  eq.value, do.log = TRUE, log.frequency = 100, verbose = TRUE,
  non.valid = "ignore", valid = function(solution) {     TRUE },
  correct = function(solution) {     solution }, resources = cresource(),
  ...)
}
\arguments{
\item{evaluate}{Function of a single parameter. Given a solution, the function returns its evaluation as a real number}

\item{initial.solution}{Solution from where the search will start}

\item{neighborhood}{Object representing the type of neighborhood to be used}

\item{cooling.scheme}{A that, given the current tempreature, returns the new (smaller) temperature}

\item{initial.temperature}{Temperature to be used at the first iteration}

\item{eq.criterion}{Criterion to be used to determine when the equilibrium for the current temperature has been reached. There are two possible values, \code{evaluations} or \code{acceptances}}

\item{eq.value}{The value associated to the equilibrium criterion. If \code{eq.cirterion = 'evaluations'} this value represents the number of evaluations. If \code{eq.cirterion = 'accepts'} this value represents the number of accepted solutions#'}

\item{do.log}{Logic value to indicate whether the progress shoul be tracked or not}

\item{log.frequency}{This value determines after how many evaluations a new log line is added. By default, its value is 100}

\item{verbose}{Logic value to indicate whether the function should print information about the search}

\item{non.valid}{Action to be performed when a non valid solution is considered. The options are \code{'ignore'}, meaning that the solution is considered anyway, \code{'discard'}, meaning that the solution is not considered and \code{'correct'}, meaning that the solution has to be corrected. This parameter has to be set only when there can be non valid solutions}

\item{valid}{A function that, given a solution, determines whether it is valid or not}

\item{correct}{A function that, given a non valid solution, corrects it. This optional parameter has to be set only with the \code{'correct'} option}

\item{resources}{Object of class \code{\link{CResource}} representing the available computational resources for the search. Bear in mind that there is no other stop criterion beyond a limited amount of resources. Therefore, you should set, at least, a limit to the total time, evaluations or iterations}

\item{...}{Special argument to pass additional parameters to the functions used in the search}
}
\value{
The function returns an object of class \code{\link{MHResult}} with all the information about the search
}
\description{
This function performs a basic simulated annealing search
}
\details{
The \code{cooling.scheme} function has to have at least one parameter, \code{temperature}, and it should produce a value smaller than the one provided in that parameter. For an example, see \code{\link{linear.cooling}}
}

