% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mutation.R
\name{swap.mutation}
\alias{swap.mutation}
\title{Swap mutation}
\usage{
swap.mutation(solution, ratio, ...)
}
\arguments{
\item{solution}{A permutation to be mutated.}

\item{ratio}{Ratio of the positions to be mutated. It has to be a value greater than 0.}
}
\description{
This function takes random pairs of positions in a permutation and swaps them
}
\examples{
id.perm <- identity.permutation(10)
swap.mutation (id.perm , ratio = 0.1)
swap.mutation (id.perm , ratio = 0.1)
swap.mutation (id.perm , ratio = 0.5)
}

