% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/selection.R
\name{tournament.selection}
\alias{tournament.selection}
\title{Tournament selection}
\usage{
tournament.selection(population, evaluation, size = 2, ...)
}
\arguments{
\item{population}{List of individuals where the selection will be applied.}

\item{evaluation}{Vector with the evaluation of the individuals.}

\item{size}{Number of individuals to be selected.}

\item{...}{Ignored}
}
\description{
This function implements the tournament selection.
}
\details{
This function takes random pairs of individuals and selects the best one
}
\examples{
url <- system.file("bays29.xml.zip", package = "metaheuR")
cost.matrix <- tsplib.parser(url)
tsp <- tsp.problem(cost.matrix)
n <- ncol(cost.matrix)
rnd.pop <- lapply(1:20 , FUN = function(x) random.permutation(n))
eval <- unlist(lapply (rnd.pop , FUN = tsp$evaluate))
tournament.selection(population = rnd.pop , evaluation = eval , size = 5)
}

