% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{tournamentSelection}
\alias{tournamentSelection}
\title{Tournament selection}
\usage{
tournamentSelection(population, evaluation, size = 2, ...)
}
\arguments{
\item{population}{List of individuals where the selection will be applied.}

\item{evaluation}{Vector with the evaluation of the individuals.}

\item{size}{Number of individuals to be selected.}

\item{...}{Ignored}
}
\description{
This function implements the tournament selection.
}
\details{
This function takes random pairs of individuals and selects the best one. If the population contains no individual, \code{NULL} is returned. If there is only one individual, it is returned \code{size} times.
}
\examples{

url <- system.file("bays29.xml.zip", package="metaheuR")
cost.matrix <- tsplibParser(url)
tsp <- tspProblem(cost.matrix)
n <- ncol(cost.matrix)
rnd.pop <- lapply(1:20, 
                  FUN=function(x) {
                        return(randomPermutation(n))
                      })
eval <- unlist(lapply(rnd.pop, FUN=tsp$evaluate))
tournamentSelection(population=rnd.pop, evaluation=eval, size=5)

}

