% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/constructive_algorithms.R
\name{tsp.greedy}
\alias{tsp.greedy}
\title{Constructive greedy algorithm for the TSP problem}
\usage{
tsp.greedy(cmatrix, cl.size = 1)
}
\arguments{
\item{cmatrix}{Cost matrix associated to the TSP instance}

\item{cl.size}{Size of candidate list.}
}
\value{
A permutation containing a solution for the problem
}
\description{
This function implements a simple constructive greedy algorithm for the TSP problem.
}
\details{
The algorithm builds the solution iteratively, selecting, at each step, the closest city to the one added in the previous step.
}
\examples{
n <- 10
cost.matrix <- matrix(runif(n^2) , ncol = n)
tsp.greedy (cost.matrix, 2)
}
\seealso{
Other Constructive algorithms: \code{\link{mis.greedy}}
}

