% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/UnivariateMarginals.R
\docType{class}
\name{UnivariateMarginals-class}
\alias{UnivariateMarginals-class}
\title{An S4 class to represent distributions based on univariate marginals}
\description{
An S4 class to represent distributions based on univariate marginals
}
\section{Slots}{

\describe{
\item{\code{prob.table}}{A matrix with the marginal probabilities. Each column is a variable and each row a possible value.}

\item{\code{levels}}{A list of levels in case the model is constructed over factors. If it is to be used with logical values, then this slot will be \code{NULL}.}
}}

