% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnivariateMarginals.R
\name{univariateMarginals}
\alias{univariateMarginals}
\title{Bais consturctor of flip neighborhoods}
\usage{
univariateMarginals(data, ...)
}
\arguments{
\item{data}{A matrix or data frame with the values to estimate the marginal probabilities. The columns represent the variables and the rwos the samples. Values can be logical, factors or characters}

\item{...}{Ignored}
}
\value{
An object of class \code{\linkS4class{UnivariateMarginals}} that includes the marginal probability distributions
}
\description{
This function creates an object of class \code{\linkS4class{FlipNeighborhood}}
}
\seealso{
Other EDA: \code{\link{bayesianNetwork}},
  \code{\link{ehbsa}}, \code{\link{mallowsModel}},
  \code{\link{plackettLuce}}
}

