% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VectorPheromone.R
\docType{class}
\name{VectorPheromone-class}
\alias{VectorPheromone-class}
\title{An S4 class to basic pheromone models for categorical vectors}
\description{
An S4 class to basic pheromone models for categorical vectors
}
\section{Slots}{

\describe{
\item{\code{trail}}{A matrix with the pheromone values. Each column is a variable and each row a possible value.}

\item{\code{binary}}{A logical value indicating whether the vector is binary or factor}

\item{\code{evaporation.factor}}{A value between 0 and 1 (both excluded) that indicates the factor used in the evaporation phase}
}}

