% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VectorPheromone.R
\name{vectorPheromone}
\alias{vectorPheromone}
\title{Bais consturctor of pheromone models for vectors}
\usage{
vectorPheromone(binary = FALSE, initial.trail, evaporation.factor, ...)
}
\arguments{
\item{initial.trail}{An square matrix containing the intial pheromone value for each link in the complete graph}

\item{evaporation.factor}{A number between 0 and 1 (both excluded) that represents the factor used in the evaporation phase}

\item{...}{Ignored}
}
\value{
An object of class \code{\linkS4class{VectorPheromone}} that includes the marginal probability distributions
}
\description{
This function creates an object of class \code{\linkS4class{VectorPheromone}}
}
\seealso{
Other ACO: \code{\link{permuLinkPheromone}};
  \code{\link{permuPosPheromone}}
}

