% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CResource.R
\name{what.finished}
\alias{what.finished}
\title{Function check what computational resources are finished (if any)}
\usage{
what.finished(resource)
}
\arguments{
\item{resource}{Object of class \code{\link{CResource}} to which the values will be substracted}
}
\value{
A vector of characters that may include \code{'time'}, \code{'evaluations'} and/or \code{'iterations'}. In case it is empty, there are still resources available
}
\description{
This function determines what (time, evaluations and/or iterations) is finished
}

