% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CResource.R
\name{whatFinished}
\alias{whatFinished}
\title{Function check what computational resources are finished (if any)}
\usage{
whatFinished(resource)
}
\arguments{
\item{resource}{Object of class \code{\linkS4class{CResource}} to which the values will be substracted}
}
\value{
A vector of characters that may include \code{'time'}, \code{'evaluations'} and/or \code{'iterations'}. In case it is empty, there are still resources available
}
\description{
This function determines what (time, evaluations and/or iterations) is finished
}

