% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CResource.R
\docType{class}
\name{CResource-class}
\alias{CResource-class}
\title{An S4 class to represent computational resources}
\description{
An S4 class to represent computational resources
}
\section{Slots}{

\describe{
\item{\code{time.total}}{A positive real number representing the total time available. It can be null, indicating that there is no time limit}

\item{\code{evaluations.total}}{A positive integer number representing the total number of evaluations available. It can be null, indicating that there is no evaluations limit}

\item{\code{iterations.total}}{A positive integer number representing the total number of iterations available. It can be null, indicating that there is no iterations limit}

\item{\code{time.consumed}}{A positive real number representing the time consumed.}

\item{\code{evaluations.consumed}}{A positive integer number representing the number of evaluations consumed.}

\item{\code{iterations.consumed}}{A positive integer number representing the number of iterations consumed.}
}}

