% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01Utils.R
\docType{data}
\name{examplegrid}
\alias{examplegrid}
\title{preprocessing combinations example}
\format{A GridClass object}
\usage{
examplegrid
}
\description{
An example grid object made of modified Iris-data with package preprocomb. \cr
Contains 540 preprocessing combinations and corresponding preprocessed data sets. \cr
Can be used with metaheur() or metaheurhyper() functions as an example grid.
}
\details{
modifiediris <- droplevels(iris[-c(1:60),]) \cr
examplegrid <- preprocomb::setgrid(phases=c("scaling", "smoothing", "outliers", "selection", "sampling"), data=modifiediris) \cr
}
\keyword{datasets}

