% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ExchangeNeighborhood.R
\docType{class}
\name{ExchangeNeighborhood-class}
\alias{ExchangeNeighborhood-class}
\title{An S4 class to represent exchange neighborhood for permutations}
\description{
An S4 class to represent exchange neighborhood for permutations
}
\section{Slots}{

\describe{
\item{\code{time_total}}{A positive real number representing the total time available. It can be null, indicating that there is no time limit}
}}

