% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/02Metaheur.R
\name{getbestheur}
\alias{getbestheur}
\title{getbestheur}
\usage{
getbestheur(x)
}
\arguments{
\item{x}{output of function metaheur()}
}
\value{
list best combination, best combinations classification accuracy
}
\description{
get the best combination and its classification accuracy
}

