% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HammingNeighborhood.R
\name{hammingNeighborhood}
\alias{hammingNeighborhood}
\title{Basic consturctor of Hamming neighborhoods}
\usage{
hammingNeighborhood(base, random = FALSE)
}
\arguments{
\item{base}{Base solution for the neighborhood. It has to be a vector of factors (all with the same levels)}

\item{random}{A logical value indicating whether the exploration should be done at random}
}
\value{
An object of class \code{\linkS4class{HammingNeighobrhood}}
}
\description{
This function creates an object of class \code{\linkS4class{HammingNeighborhood}}
}
\seealso{
\code{\link{hasMoreNeighbors}} \code{\link{resetNeighborhood}} \code{\link{nextNeighbor}}

Other neighborhoods: \code{\link{exchangeNeighborhood}},
  \code{\link{flipNeighborhood}},
  \code{\link{hasMoreNeighbors}},
  \code{\link{nextNeighbor}},
  \code{\link{resetNeighborhood}},
  \code{\link{swapNeighborhood}}
}

