% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/permutation_problems.R
\name{lopProblem}
\alias{lopProblem}
\title{LOP problem evaluator}
\usage{
lopProblem(matrix)
}
\arguments{
\item{matrix}{matrix for the LOP problem}
}
\value{
A function that can be used to evaluate solutions for a LOP problem
}
\description{
This function generates an evaluation function associated
with a LOP problem
}
\examples{
matrix <- matrix(runif(100), ncol=10)
lop <- lopProblem(matrix)
lop$evaluate(randomPermutation(10))
}

