% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MallowsModel.R
\name{mallowsModel}
\alias{mallowsModel}
\title{Bais consturctor of flip neighborhoods}
\usage{
mallowsModel(data, distance = "Kendall", ...)
}
\arguments{
\item{data}{A matrix or data frame with the values to estimate the marginal probabilities. The columns represent the variables and the rwos the samples. Values can be logical, factors or characters}

\item{...}{Ignored}
}
\value{
An object of class \code{\linkS4class{UnivariateMarginals}} that includes the marginal probability distributions
}
\description{
This function creates an object of class \code{\linkS4class{FlipNeighborhood}}
}
\seealso{
Other EDA: \code{\link{univariateMarginals}}
}

