% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHResult.R
\name{mHResult}
\alias{mHResult}
\title{Function to construct objects of type \code{mHResult}}
\usage{
mHResult(algorithm, parameters, solution, evaluation, resources,
  description = "", log = NULL, solutions = NULL)
}
\arguments{
\item{algorithm}{String indicating the algorithm run}

\item{parameters}{List of parameters used in the experiment}

\item{solution}{Best found solution}

\item{evaluation}{Optimal value of the objective function}

\item{resources}{Object of class \code{\link{CResource}} contining the information about the use of resources in the experiment}

\item{description}{Description of the algorithm run}

\item{log}{Matrix with the evolution of the search}
}
\value{
A new object of class \code{\link{mHResult}} with the updated remaining resources
}
\description{
This function creates objects of class \code{\link{mHResult}}
}

