% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/03MetaheurCompare.R
\name{plotsearchpath}
\alias{plotsearchpath}
\title{plotsearchpath}
\usage{
plotsearchpath(x)
}
\arguments{
\item{x}{output of metaheurcompare function}
}
\description{
plot the searh path of metaheurcompare()
}
\examples{
## result <- metaheurcompare()
## plotsearchpath(result)
}

