% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06ExtractComponent.R
\name{plotsearchpath}
\alias{plotsearchpath}
\title{plot search path}
\usage{
plotsearchpath(x)
}
\arguments{
\item{x}{(metaheur object or list of metaheurgrid class objects) object to be plotted}
}
\description{
Plots the search path for metaheur() and three scenarios in plotting
the result of metaheurgrid()
}
\examples{
plotsearchpath(examplesearch)
plotsearchpath(examplehyperparam)
}

