% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CResource.R
\name{remaining.evaluations}
\alias{remaining.evaluations}
\title{Function to get the remaining evaluations available}
\usage{
remaining.evaluations(resource)
}
\arguments{
\item{resource}{Object of class \code{\linkS4class{CResource}} whose information we are quering}
}
\value{
Remaining number of evaluations
}
\description{
This function returns the remaining evaluations available
}

