% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CResource.R
\name{remaining.iterations}
\alias{remaining.iterations}
\title{Function to get the remaining iterations available}
\usage{
remaining.iterations(resource)
}
\arguments{
\item{resource}{Object of class \code{\linkS4class{CResource}} whose information we are quering}
}
\value{
Remaining time
}
\description{
This function returns the remaining iterations available
}

