% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SwapNeighborhood.R
\name{swapNeighborhood}
\alias{swapNeighborhood}
\title{Bais consturctor of swap neighborhoods}
\usage{
swapNeighborhood(base, random = FALSE)
}
\arguments{
\item{base}{Base solution for the neighborhood. It has to be an object of class \code{\linkS4class{Permutation}}}

\item{random}{A logical value indicating whether the exploration should be done at random}
}
\value{
An object of class \code{\linkS4class{SwapNeighborhood}}
}
\description{
This function creates an object of class \code{\linkS4class{SwapNeighborhood}}
}
\seealso{
\code{\link{hasMoreNeighbors}} \code{\link{resetNeighborhood}} \code{\link{nextNeighbor}}

Other neighborhoods: \code{\link{flipNeighborhood}};
  \code{\link{hammingNeighborhood}},
  \code{\link{hammingNeighborhood}};
  \code{\link{has.more.neighbors}};
  \code{\link{next.neighbor}};
  \code{\link{reset.neighborhood}}
}

