% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/permutation_problems.R
\name{tspProblem}
\alias{tspProblem}
\title{TSP problem evaluator}
\usage{
tspProblem(cmatrix)
}
\arguments{
\item{cmatrix}{Cost matrix for the TSP problem}
}
\value{
A function that can be used to evaluate solutions for a TSP problem
}
\description{
This function generates an evaluation function associated
with a TSP problem
}
\examples{
cmatrix <- matrix(runif(100), ncol=10)
tsp <- tspProblem(cmatrix)
tsp$evaluate(randomPermutation(10))
}

